pyproject.toml
[tool.poetry]
name = "groceries_scraper_package"
version = "0.1.0"
description = ""
authors = ["Warnalytics <warner.d.jong@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9.6"
pyleniumio = "^1.15.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
